add toggler for container in modal for edit data birthday
